move purge_data.py and accessdb.py inside pybossa folder (#797)
* move purge_data.py and accessdb.py inside pybossa folder

* fix unit test
