+++ test/test_jobs/__init__.py
@@ -24,8 +24,6 @@ from unittest.mock import patch
 from nose.tools import assert_raises
 from test import with_context, Test
 from test.factories import ProjectFactory, TaskFactory, TaskRunFactory
-from accessdb import AccessDatabase
-from sqlalchemy.sql import text
 
 
 def jobs():
@@ -252,7 +250,7 @@ class TestJobs(Test):
             assert job['queue'] == "weekly"
 
     @with_context
-    @patch('pybossa.jobs.purge_task_data')
+    @patch('pybossa.purge_data.purge_task_data')
     def test_completed_tasks_cleanup(self, mock_purge_tasks):
         """Test completed_tasks_cleanup deletes tasks qualify for deletion."""
 
@@ -301,7 +299,7 @@ class TestJobs(Test):
         assert saturday.strftime("%Y-%m-%d %H:%M:%S") == "2026-01-31 16:00:00"
 
     @with_context
-    @patch('pybossa.jobs.purge_task_data')
+    @patch('pybossa.purge_data.purge_task_data')
     def test_completed_tasks_cleanup_bad_config(self, mock_purge_tasks):
         """Test completed_tasks_cleanup deletes tasks qualify for deletion."""
 
@@ -311,7 +309,7 @@ class TestJobs(Test):
         assert not mock_purge_tasks.called
 
     @with_context
-    @patch('pybossa.jobs.purge_task_data')
+    @patch('pybossa.purge_data.purge_task_data')
     def test_completed_tasks_cleanup_bad_project_config(self, mock_purge_tasks):
         """Test completed_tasks_cleanup deletes tasks qualify for deletion."""
 
@@ -327,7 +325,7 @@ class TestJobs(Test):
     # archive tables from test db upon testing complete for future test runs to be successful
     # mock_purge_tasks can be removed with task data cleanup and archive happening in actual
     # @with_context
-    # @patch('pybossa.jobs.purge_task_data')
+    # @patch('pybossa.purge_data.purge_task_data')
     # def test_completed_tasks_cleanup(self, mock_purge_tasks):
     #     """Test completed_tasks_cleanup deletes tasks qualify for deletion."""
